requirements.txt
google-generativeai
moviepy
python-dotenv
Pillow
kokoro>=0.9.4
soundfile
